refactor: App 컴포넌트 위치 변경 (components → src)
