+++ src/Main.jsx
@@ -1,5 +1,5 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import App from "./components/App";
+import App from "./App";
 import ItemsPage from "./pages/ItemsPage";
 import BoardPage from "./pages/BoardPage";
 import AddItemPage from "./pages/AddItemPage";
